readCSV.c is now reading input data in CSR format

+++ CPU Implementation/validation.c
@@ -8,17 +8,17 @@
  *
  **********************************************************************/
 
-#include "validation.h"
 #include <stdio.h>
 #include <stdlib.h>
 #include <math.h>
+#include "validation.h"
 
-int validation(int nT, int nT_Mat)
+int validation(long nT, long nT_Mat)
 {
 
   if ( nT != nT_Mat )
   {
-    printf("Validation FAILED: nT = %d, while correct value nT_Mat = %d \n", nT, nT_Mat);
+    printf("Validation FAILED: nT = %ld, while correct value nT_Mat = %ld \n", nT, nT_Mat);
     return 0;
   }
